@@ -1,6 +1,6 @@
 // theme.ts
-export const textColorLight = "#404067";
-export const textColorDark = "#D3CDF9";
+export const textColorLight = '#404067';
+export const textColorDark = '#D3CDF9';
 
-export const formBackgroundLight = "#D3CDF9";
-export const formBackgroundDark = "#6B60B1";
+export const formBackgroundLight = '#D3CDF9';
+export const formBackgroundDark = '#6B60B1';
