@@ -1,13 +1,13 @@
 // pages/_app.js
-import { ChakraProvider } from "@chakra-ui/react";
-import { AppProps } from "next/app"; // Import the AppProps type
+import { ChakraProvider } from '@chakra-ui/react';
+import { AppProps } from 'next/app'; // Import the AppProps type
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return (
-    <ChakraProvider>
-      <Component {...pageProps} />
-    </ChakraProvider>
-  );
+	return (
+		<ChakraProvider>
+			<Component {...pageProps} />
+		</ChakraProvider>
+	);
 }
 
 export default MyApp;
